# Only update VCS when things have actually changed