--- a/pip/vcs/git.py
+++ b/pip/vcs/git.py
@@ -65,7 +65,7 @@ class Git(VersionControl):
         and branches may need origin/ as a prefix.
         Returns the SHA1 of the branch or tag if found.
         """
-        revisions = self.get_refs(dest)
+        revisions = self.get_short_refs(dest)
 
         origin_rev = 'origin/%s' % rev
         if origin_rev in revisions:
@@ -79,6 +79,15 @@ class Git(VersionControl):
                 "Could not find a tag or branch '%s', assuming commit.", rev,
             )
             return rev_options
+
+    def check_version(self, dest, rev_options):
+        """
+        Compare the current sha to the ref. ref may be a branch or tag name,
+        but current rev will always point to a sha. This means that a branch
+        or tag will never compare as True. So this ultimately only matches
+        against exact shas.
+        """
+        return self.get_revision(dest).startswith(rev_options[0])
 
     def switch(self, dest, url, rev_options):
         self.run_command(['config', 'remote.origin.url', url], cwd=dest)
@@ -115,7 +124,7 @@ class Git(VersionControl):
             if rev:
                 rev_options = self.check_rev_options(rev, dest, rev_options)
                 # Only do a checkout if rev_options differs from HEAD
-                if not self.get_revision(dest).startswith(rev_options[0]):
+                if not self.check_version(dest, rev_options):
                     self.run_command(
                         ['checkout', '-q'] + rev_options,
                         cwd=dest,
@@ -134,23 +143,48 @@ class Git(VersionControl):
             ['rev-parse', 'HEAD'], show_stdout=False, cwd=location)
         return current_rev.strip()
 
-    def get_refs(self, location):
-        """Return map of named refs (branches or tags) to commit hashes."""
+    def get_full_refs(self, location):
+        """Yields tuples of (commit, ref) for branches and tags"""
         output = self.run_command(['show-ref'],
                                   show_stdout=False, cwd=location)
-        rv = {}
         for line in output.strip().splitlines():
             commit, ref = line.split(' ', 1)
-            ref = ref.strip()
+            yield commit.strip(), ref.strip()
+
+    def is_ref_remote(self, ref):
+        return ref.startswith('refs/remotes/')
+
+    def is_ref_branch(self, ref):
+        return ref.startswith('refs/heads/')
+
+    def is_ref_tag(self, ref):
+        return ref.startswith('refs/tags/')
+
+    def is_ref_commit(self, ref):
+        """A ref is a commit sha if it is not anything else"""
+        return not any((
+            self.is_ref_remote(ref),
+            self.is_ref_branch(ref),
+            self.is_ref_tag(ref),
+        ))
+
+    # Should deprecate `get_refs` since it's ambiguous
+    def get_refs(self, location):
+        return self.get_short_refs(location)
+
+    def get_short_refs(self, location):
+        """Return map of named refs (branches or tags) to commit hashes."""
+        rv = {}
+        for commit, ref in self.get_full_refs(location):
             ref_name = None
-            if ref.startswith('refs/remotes/'):
+            if self.is_ref_remote(ref):
                 ref_name = ref[len('refs/remotes/'):]
-            elif ref.startswith('refs/heads/'):
+            elif self.is_ref_branch(ref):
                 ref_name = ref[len('refs/heads/'):]
-            elif ref.startswith('refs/tags/'):
+            elif self.is_ref_tag(ref):
                 ref_name = ref[len('refs/tags/'):]
             if ref_name is not None:
-                rv[ref_name] = commit.strip()
+                rv[ref_name] = commit
         return rv
 
     def get_src_requirement(self, dist, location, find_tags):
@@ -161,7 +195,7 @@ class Git(VersionControl):
         if not repo:
             return None
         current_rev = self.get_revision(location)
-        refs = self.get_refs(location)
+        refs = self.get_short_refs(location)
         # refs maps names to commit hashes; we need the inverse
         # if multiple names map to a single commit, we pick the first one
         # alphabetically