--- a/tests/unit/test_vcs.py
+++ b/tests/unit/test_vcs.py
@@ -1,3 +1,4 @@
+import pytest
 from tests.lib import pyversion
 from pip.vcs import VersionControl
 from pip.vcs.bazaar import Bazaar
@@ -10,7 +11,8 @@ else:
     VERBOSE_FALSE = 0
 
 
-def test_git_get_src_requirements():
+@pytest.fixture
+def git():
     git_url = 'http://github.com/pypa/pip-test-package'
     refs = {
         '0.1': 'a8992fc7ee17e5b9ece022417b64594423caca7c',
@@ -27,9 +29,18 @@ def test_git_get_src_requirements():
     git = Git()
     git.get_url = Mock(return_value=git_url)
     git.get_revision = Mock(return_value=sha)
-    git.get_refs = Mock(return_value=refs)
+    git.get_short_refs = Mock(return_value=refs)
+    return git
+
+
+@pytest.fixture
+def dist():
     dist = Mock()
     dist.egg_name = Mock(return_value='pip_test_package')
+    return dist
+
+
+def test_git_get_src_requirements(git, dist):
     ret = git.get_src_requirement(dist, location='.', find_tags=None)
 
     assert ret == ''.join([
@@ -37,6 +48,16 @@ def test_git_get_src_requirements():
         '@5547fa909e83df8bd743d3978d6667497983a4b7',
         '#egg=pip_test_package-bar'
     ])
+
+
+@pytest.mark.parametrize('ref,result', (
+    ('5547fa909e83df8bd743d3978d6667497983a4b7', True),
+    ('5547fa909', True),
+    ('abc123', False),
+    ('foo', False),
+))
+def test_git_check_version(git, ref, result):
+    assert git.check_version('foo', ref) is result
 
 
 def test_translate_egg_surname():