@dcramer 2015-08-31
This will help us out greatly, as the git checks take a majority of the time on a no-op install:

```
dcramer:getsentry in ~/Development/getsentry on git:staging
$ less requirements.txt | grep git+ | wc -l
      30
```